+++ src/api/routes/v1/auth.route.ts
@@ -3,12 +3,7 @@ const express = require('express');
 const validate = require('express-validation');
 const controller = require('../../controllers/auth.controller');
 const oAuthLogin = require('../../middlewares/auth').oAuth;
-const {
-  login,
-  register,
-  oAuth,
-  refresh,
-} = require('../../validations/auth.validation');
+const { login, register, oAuth, refresh, forgotPassword } = require('../../validations/auth.validation');
 
 const router = express.Router();
 
@@ -39,9 +34,7 @@ const router = express.Router();
  *
  * @apiError (Bad Request 400)  ValidationError  Some parameters may contain invalid values
  */
-router.route('/register')
-  .post(validate(register), controller.register);
-
+router.route('/register').post(validate(register), controller.register);
 
 /**
  * @api {post} v1/auth/login Login
@@ -70,9 +63,7 @@ router.route('/register')
  * @apiError (Bad Request 400)  ValidationError  Some parameters may contain invalid values
  * @apiError (Unauthorized 401)  Unauthorized     Incorrect email or password
  */
-router.route('/login')
-  .post(validate(login), controller.login);
-
+router.route('/login').post(validate(login), controller.login);
 
 /**
  * @api {post} v1/auth/refresh-token Refresh Token
@@ -93,14 +84,11 @@ router.route('/login')
  * @apiError (Bad Request 400)  ValidationError  Some parameters may contain invalid values
  * @apiError (Unauthorized 401)  Unauthorized     Incorrect email or refreshToken
  */
-router.route('/refresh-token')
-  .post(validate(refresh), controller.refresh);
-
+router.route('/refresh-token').post(validate(refresh), controller.refresh);
 
 /**
  * TODO: POST /v1/auth/reset-password
  */
-
 
 /**
  * @api {post} v1/auth/facebook Facebook Login
@@ -120,8 +108,7 @@ router.route('/refresh-token')
  * @apiError (Bad Request 400)  ValidationError  Some parameters may contain invalid values
  * @apiError (Unauthorized 401)  Unauthorized    Incorrect access_token
  */
-router.route('/facebook')
-  .post(validate(oAuth), oAuthLogin('facebook'), controller.oAuth);
+router.route('/facebook').post(validate(oAuth), oAuthLogin('facebook'), controller.oAuth);
 
 /**
  * @api {post} v1/auth/google Google Login
@@ -141,8 +128,8 @@ router.route('/facebook')
  * @apiError (Bad Request 400)  ValidationError  Some parameters may contain invalid values
  * @apiError (Unauthorized 401)  Unauthorized    Incorrect access_token
  */
-router.route('/google')
-  .post(validate(oAuth), oAuthLogin('google'), controller.oAuth);
+router.route('/google').post(validate(oAuth), oAuthLogin('google'), controller.oAuth);
 
+router.route('/forgot-password').post(validate(forgotPassword), controller.forgotPassword);
 
 module.exports = router;