+++ src/api/utils/Utils.ts
@@ -174,3 +174,8 @@ export async function apiJson({ req, res, data, model, meta = {}, json = false }
   }
   return res.json(output);
 }
+
+export function randomString(len = 10, charStr = 'abcdefghijklmnopqrstuvwxyz0123456789') {
+  const chars = [...(charStr + '')];
+  return [...Array(len)].map(i => chars[(Math.random() * chars.length) | 0]).join('');
+}