+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [0.6.1] - 2019-02-27
+## [0.6.5] - 2019-02-28
 
 ### Added
 - support for "&fields" param in Model.transform(req) to include certain fields in API response
@@ -8,9 +8,11 @@
 - added MsgUtils sendEmail (using nodemailer & mailgun)
 - added MsgUtils email template function, e.g. sendEmail(welcomeEmail({ name, email }))
 - added multer to handle file upload
-- added "features.md" to explain features in details.
+- added "features.md" to explain features in details
+- added /forgot-password route & controller
 ### Fixed
 - fixed yarn lint
+- fixed lint errors
 
 ## [0.4.7] - 2019-02-21